+++ RtspAnalyser.Motion/includes/DelNico/RtspAnalyser/Motion/MotionManagerCalling.h
@@ -8,16 +8,16 @@ namespace DelNico::RtspAnalyser::Motion {
             MotionManagerCalling();
             MotionManagerCalling(MotionManagerCaller who);
             MotionManagerCalling(MotionManagerCaller who, bool human_detected);
-            MotionManagerCalling(MotionManagerCaller who, bool human_detected, float score);
+            MotionManagerCalling(MotionManagerCaller who, bool human_detected, int score);
             ~MotionManagerCalling();
 
             MotionManagerCaller getWho() const;
             bool isHumanDetected() const;
-            float getScore() const;
+            int getScore() const;
         private:
             MotionManagerCaller who;
             bool human_detected;
-            float score;
+            int score;
 
             const bool operator==(const MotionManagerCalling & other) const;
     };
